pyproject.toml
 [tool.poetry]
 name = "fastapi-clean-arch"
 version = "0.1.0"
 description = "FastAPI application following a clean architecture with method-per-file structure."
authors = ["Akhmedov Jacob <jecob.akhmedov@gmail.com>"]
homepage = "https://github.com/Jacobamv"
repository = "https://github.com/Jacobamv/fastapi-clean-arch"
 readme = "README.md"
packages = [
    { include = "app" },
]

 [tool.poetry.dependencies]
 python = "^3.11"
fastapi = "^0.121.1"
uvicorn = { extras = ["standard"], version = "^0.38.0" }
sqlalchemy = "^2.0.44"
alembic = "^1.17.1"
python-dotenv = "^1.2.1"
psycopg2-binary = "^2.9.11"
python-jose = { extras = ["cryptography"], version = "^3.5.0" }
passlib = "^1.7.4"
bcrypt = "3.2.0"
pydantic = "^2.12.4"
pydantic-settings = "^2.11.0"
 email-validator = "^2.3.0"

 [tool.poetry.group.dev.dependencies]
pytest = "^9.0.0"
ruff = "^0.14.4"

 [tool.ruff]
 src = ["app"]
 line-length = 100

 [build-system]
 requires = ["poetry-core>=1.5.0"]
 build-backend = "poetry.core.masonry.api"

